+++ src/femto/rasterimage.py
@@ -86,11 +86,11 @@ class RasterImage(LaserPath):
 
             # add a path for each sub-split with open shutter
             for x_split in x_open_shutter:
-                x_row = np.array([x_split[0], x_split[0], x_split[-1], x_split[-1], x_split[0]], dtype=np.float32)
-                y_row = y_val * np.ones_like(x_row, dtype=np.float32)
-                z_row = z_val * np.ones_like(x_row, dtype=np.float32)
+                x_row = np.array([x_split[0], x_split[0], x_split[-1], x_split[-1], x_split[0]], dtype=np.float64)
+                y_row = y_val * np.ones_like(x_row, dtype=np.float64)
+                z_row = z_val * np.ones_like(x_row, dtype=np.float64)
                 f_row = np.array(
-                    [self.speed_closed, self.speed, self.speed, self.speed_closed, self.speed_closed], dtype=np.float32
+                    [self.speed_closed, self.speed, self.speed, self.speed_closed, self.speed_closed], dtype=np.float64
                 )
                 s_row = np.array([0, 1, 1, 0, 0], dtype=int)
